Restructures ProductCard and ProductList
