+++ src/components/ProductList/ProductCard/ProductCard.jsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
-import Button from "../ui/Button";
-import Rating from "../ui/Rating";
-import { isOnSale } from "../../utils";
+import Button from "../../ui/Button";
+import Rating from "../../ui/Rating";
+import { isOnSale } from "../../../utils";
 import {
   ProductContainer,
   PriceAndButton,
