+++ src/components/ProductList/ProductList.jsx
@@ -1,4 +1,4 @@
-import ProductCard from "../ProductCard";
+import ProductCard from "./ProductCard";
 import { ProductGrid } from "./ProductList.styled";
 
 const ProductList = ({ products }) => {
